minor updates to bl_info and updater verbiage; allow all addon files to be updated
Added proper github URL, name, author, desc, and category to bl_info

Updated the addon updater to use the actual human-readable name of the addon in places where it normally is visible to the user. The updater requires you have a "name" that is lower case and no spaces, and then shows that restricted name to the user, which just looks bad. Since the addon_updater_ops.py is intended to be edited and customized by the addon developer anyhow, it seemed reasonable to put the addon's actual human-readable name in places where humans would normally see it.

Also set the updater to allow all files to be updated, as the existing list wouldn't allow *.blend and *.json files to be updated, which is almost certainly not what we want.