obj_import: don't error on duplicate faces (probably for cloaks/etc)
sometimes there are duplicate faces when importing, which right now
makes the importer bail out. Spot checking these, the duplicates
tend to be the same as a previous face, except with a vert order
of (2,1,3) instead of (1,2,3), which gives the duplicate face the
opposite normal of the one it is duplicating. We're pretty sure
these are used for cloaks and other double-sided things, since the
WoW engine doesn't believe in double-sided polys. There may be some
situations where there's something different going on, and we'd
really like to find/investigate those if they exist, but for now,
just ignoring duplicate faces will stop the addon from crashing, with
no apparent downsides.

This may be a problem if there's some object in WoW where the opposite
sides have different textures (which throwing away one of the sides in
blender will cause it to be rendered incorrectly), but who knows if
that exists. Can deal with it if we find it.