use mashumaro for type checking/access to metadata json data
mashumaro is very very very slightly slower than using a raw dict
or a TypedDict, but gives full type annotations and attribute-like
access to data from the metadata json, as well as providing default
values and other such. Probably still needs some adjustments, but
seems to work well.

We vendor it because for some reason blender wouldn't find it
properly in some circumstances if it was installed in other ways.
There's now a 'vendor' subdirectory and a requirements.txt (and
associated requirements.in) for populating it. Run 'make vendor'
to populate it. You will need the pip-compile-multi and pdistx
packages installed from pypi to manage your vendor directory.

There's a bug in mashumaro that breaks it when vendored, which
is patched by the Makefile until the bug is fixed upstream (an
issue has been opened for it.